Fix flaky bucket eviction tests by splitting/isolating them (#5876)
* fix flaky bucket eviction tests by splitting/isolating them

* fix imports
