+++ frontend/javascripts/test/sagas/volumetracing/volumetracing_saga_integration.spec.js
@@ -1,10 +1,8 @@
 // @flow
 /* eslint-disable no-await-in-loop */
 
-import mockRequire from "mock-require";
 import test from "ava";
-import { waitForCondition } from "libs/utils";
-import _ from "lodash";
+import mockRequire from "mock-require";
 
 import "test/sagas/saga_integration.mock";
 import {
@@ -489,98 +487,3 @@ test.serial("Brushing/Tracing with undo (II)", async t => {
   t.is(await t.context.api.data.getDataValue("segmentation", [1, 0, 0]), newCellId + 1);
   t.is(await t.context.api.data.getDataValue("segmentation", [5, 0, 0]), oldCellId);
 });
-
-test.serial(
-  "Brushing/Tracing should crash when too many buckets are labeled at once without saving inbetween",
-  async t => {
-    await t.context.api.tracing.save();
-
-    t.context.mocks.Request.sendJSONReceiveArraybufferWithHeaders = createBucketResponseFunction(
-      Uint16Array,
-      0,
-      0,
-    );
-    // webKnossos will start to evict buckets forcefully if too many are dirty at the same time.
-    // This is not ideal, but usually handled by the fact that buckets are regularly saved to the
-    // backend and then marked as not dirty.
-    // This test provokes that webKnossos crashes (a hard crash is only done during testing; in dev/prod
-    // a soft warning is emitted via the devtools).
-    // The corresponding sibling test checks that saving inbetween does not make webKnossos crash.
-    t.plan(2);
-    t.false(hasRootSagaCrashed());
-    const failedSagaPromise = waitForCondition(hasRootSagaCrashed, 500);
-    await Promise.race([testLabelingManyBuckets(t, false), failedSagaPromise]);
-    t.true(hasRootSagaCrashed());
-  },
-);
-
-test.serial(
-  "Brushing/Tracing should not crash when too many buckets are labeled at once with saving inbetween",
-  async t => {
-    await t.context.api.tracing.save();
-
-    t.context.mocks.Request.sendJSONReceiveArraybufferWithHeaders = createBucketResponseFunction(
-      Uint16Array,
-      0,
-      0,
-    );
-    t.plan(2);
-    t.false(hasRootSagaCrashed());
-    const failedSagaPromise = waitForCondition(hasRootSagaCrashed, 500);
-    await Promise.race([testLabelingManyBuckets(t, true), failedSagaPromise]);
-    t.false(hasRootSagaCrashed());
-  },
-);
-
-async function testLabelingManyBuckets(t, saveInbetween) {
-  // We set MAXIMUM_BUCKET_COUNT to 150 and then label 199 = 75 (mag1) + 124 (downsampled) buckets in total.
-  // In between, we will save the data which allows the buckets of the first batch to be GC'ed.
-  // Therefore, saving the buckets of the second batch should not cause any problems.
-  t.context.model.getCubeByLayerName("segmentation").MAXIMUM_BUCKET_COUNT = 150;
-
-  const oldCellId = 11;
-  const brushSize = 10;
-  const newCellId = 2;
-
-  t.context.mocks.Request.sendJSONReceiveArraybufferWithHeaders = createBucketResponseFunction(
-    Uint16Array,
-    oldCellId,
-    500,
-  );
-  // Reload buckets which might have already been loaded before swapping the sendJSONReceiveArraybufferWithHeaders
-  // function.
-  await t.context.api.data.reloadAllBuckets();
-
-  // Prepare to paint into the center of 50 buckets.
-  const paintPositions1 = _.range(0, 50).map(idx => [32 * idx + 16, 32 * idx + 16, 32 * idx + 16]);
-  // Prepare to paint into the center of 50 other buckets.
-  const paintPositions2 = _.range(50, 100).map(idx => [
-    32 * idx + 16,
-    32 * idx + 16,
-    32 * idx + 16,
-  ]);
-
-  Store.dispatch(updateUserSettingAction("brushSize", brushSize));
-  Store.dispatch(setToolAction(AnnotationToolEnum.BRUSH));
-  Store.dispatch(setActiveCellAction(newCellId));
-
-  for (const paintPosition of paintPositions1) {
-    Store.dispatch(setPositionAction(paintPosition));
-    Store.dispatch(startEditingAction(paintPosition, OrthoViews.PLANE_XY));
-    Store.dispatch(addToLayerAction(paintPosition));
-    Store.dispatch(finishEditingAction());
-  }
-
-  if (saveInbetween) {
-    await t.context.api.tracing.save();
-  }
-
-  for (const paintPosition of paintPositions2) {
-    Store.dispatch(setPositionAction(paintPosition));
-    Store.dispatch(startEditingAction(paintPosition, OrthoViews.PLANE_XY));
-    Store.dispatch(addToLayerAction(paintPosition));
-    Store.dispatch(finishEditingAction());
-  }
-
-  await t.context.api.tracing.save();
-}
